fix(app): move NFC-debug endpoints out of consts/links
links.ts is contract-tested to hold only https URLs with Url/Uri-style
names; the bare nfcDebugHost and wss relay URL broke that in CI. They
now live in consts/nfcDebug.ts with the host no longer exported.
